move spells/nonspells over to their own files, and retab the whole directory
