+++ src/stages/stage5/background_anim.c
@@ -17,34 +17,34 @@
 void stage5_update(void) {
 	Stage5DrawData *stage5_draw_data = stage5_get_draw_data();
 
-    TIMER(&global.timer);
-    float w = 0.005;
+	TIMER(&global.timer);
+	float w = 0.005;
 
-    stage5_draw_data->stairs.rotshift += stage5_draw_data->stairs.omega;
-    stage_3d_context.crot[0] += stage5_draw_data->stairs.omega*0.5;
-    stage5_draw_data->stairs.rad += stage5_draw_data->stairs.omega*20;
+	stage5_draw_data->stairs.rotshift += stage5_draw_data->stairs.omega;
+	stage_3d_context.crot[0] += stage5_draw_data->stairs.omega*0.5;
+	stage5_draw_data->stairs.rad += stage5_draw_data->stairs.omega*20;
 
-    int rot_time = 6350;
+	int rot_time = 6350;
 
-    FROM_TO(rot_time, rot_time+50, 1) {
-        stage5_draw_data->stairs.omega -= 0.005;
+	FROM_TO(rot_time, rot_time+50, 1) {
+		stage5_draw_data->stairs.omega -= 0.005;
 	}
 
-    FROM_TO(rot_time+200, rot_time+250, 1) {
-        stage5_draw_data->stairs.omega += 0.005;
+	FROM_TO(rot_time+200, rot_time+250, 1) {
+		stage5_draw_data->stairs.omega += 0.005;
 	}
 
-    stage_3d_context.cx[0] = stage5_draw_data->stairs.rad*cos(-w*global.frames);
-    stage_3d_context.cx[1] = stage5_draw_data->stairs.rad*sin(-w*global.frames);
-    stage_3d_context.cx[2] = -1700+w*3000/M_PI*global.frames;
+	stage_3d_context.cx[0] = stage5_draw_data->stairs.rad*cos(-w*global.frames);
+	stage_3d_context.cx[1] = stage5_draw_data->stairs.rad*sin(-w*global.frames);
+	stage_3d_context.cx[2] = -1700+w*3000/M_PI*global.frames;
 
-    stage_3d_context.crot[2] = stage5_draw_data->stairs.rotshift-180/M_PI*w*global.frames;
+	stage_3d_context.crot[2] = stage5_draw_data->stairs.rotshift-180/M_PI*w*global.frames;
 
-    stage5_draw_data->stairs.light_strength *= 0.98;
+	stage5_draw_data->stairs.light_strength *= 0.98;
 
-    if (frand() < 0.01) {
-        stage5_draw_data->stairs.light_strength = 5+5*frand();
+	if (frand() < 0.01) {
+		stage5_draw_data->stairs.light_strength = 5+5*frand();
 	}
 
-    stage3d_update(&stage_3d_context);
+	stage3d_update(&stage_3d_context);
 }
