+++ src/stages/stage5/draw.c
@@ -23,12 +23,12 @@ Stage5DrawData *stage5_get_draw_data(void) {
 
 void stage5_drawsys_init(void) {
 	stage5_draw_data = calloc(1, sizeof(*stage5_draw_data));
-    stage3d_init(&stage_3d_context, 16);
+	stage3d_init(&stage_3d_context, 16);
 
-    stage_3d_context.crot[0] = 60;
+	stage_3d_context.crot[0] = 60;
 
-    stage5_draw_data->stairs.rotshift = 140;
-    stage5_draw_data->stairs.rad = 2800;
+	stage5_draw_data->stairs.rotshift = 140;
+	stage5_draw_data->stairs.rad = 2800;
 }
 
 void stage5_drawsys_shutdown(void) {
@@ -38,52 +38,52 @@ void stage5_drawsys_shutdown(void) {
 }
 
 static uint stage5_stairs_pos(Stage3D *s3d, vec3 pos, float maxrange) {
-    vec3 p = {0, 0, 0};
-    vec3 r = {0, 0, 6000};
+	vec3 p = {0, 0, 0};
+	vec3 r = {0, 0, 6000};
 
-    return linear3dpos(s3d, pos, maxrange, p, r);
+	return linear3dpos(s3d, pos, maxrange, p, r);
 }
 
 static void stage5_stairs_draw(vec3 pos) {
-    r_state_push();
-    r_mat_mv_push();
-    r_mat_mv_translate(pos[0], pos[1], pos[2]);
-    r_mat_mv_scale(300,300,300);
-    r_shader("tower_light");
-    r_uniform_sampler("tex", "stage5/tower");
-    r_uniform_vec3("lightvec", 0, 0, 0);
-    r_uniform_vec4("color", 0.1, 0.1, 0.5, 1);
-    r_uniform_float("strength", stage5_draw_data->stairs.light_strength);
-    r_draw_model("tower");
-    r_mat_mv_pop();
-    r_state_pop();
+	r_state_push();
+	r_mat_mv_push();
+	r_mat_mv_translate(pos[0], pos[1], pos[2]);
+	r_mat_mv_scale(300,300,300);
+	r_shader("tower_light");
+	r_uniform_sampler("tex", "stage5/tower");
+	r_uniform_vec3("lightvec", 0, 0, 0);
+	r_uniform_vec4("color", 0.1, 0.1, 0.5, 1);
+	r_uniform_float("strength", stage5_draw_data->stairs.light_strength);
+	r_draw_model("tower");
+	r_mat_mv_pop();
+	r_state_pop();
 }
 
 void stage5_draw(void) {
-    r_mat_proj_perspective(STAGE3D_DEFAULT_FOVY, STAGE3D_DEFAULT_ASPECT, 100, 20000);
-    stage3d_draw(&stage_3d_context, 30000, 1, (Stage3DSegment[]) { stage5_stairs_draw, stage5_stairs_pos });
+	r_mat_proj_perspective(STAGE3D_DEFAULT_FOVY, STAGE3D_DEFAULT_ASPECT, 100, 20000);
+	stage3d_draw(&stage_3d_context, 30000, 1, (Stage3DSegment[]) { stage5_stairs_draw, stage5_stairs_pos });
 }
 
 void iku_spell_bg(Boss *b, int t) {
-    fill_viewport(0, 300, 1, "stage5/spell_bg");
+	fill_viewport(0, 300, 1, "stage5/spell_bg");
 
-    r_blend(BLEND_MOD);
-    fill_viewport(0, t*0.001, 0.7, "stage5/noise");
-    r_blend(BLEND_PREMUL_ALPHA);
+	r_blend(BLEND_MOD);
+	fill_viewport(0, t*0.001, 0.7, "stage5/noise");
+	r_blend(BLEND_PREMUL_ALPHA);
 
-    r_mat_mv_push();
-    r_mat_mv_translate(0, -100, 0);
-    fill_viewport(t/100.0, 0, 0.5, "stage5/spell_clouds");
-    r_mat_mv_translate(0, 100, 0);
-    fill_viewport(t/100.0 * 0.75, 0, 0.6, "stage5/spell_clouds");
-    r_mat_mv_translate(0, 100, 0);
-    fill_viewport(t/100.0 * 0.5, 0, 0.7, "stage5/spell_clouds");
-    r_mat_mv_translate(0, 100, 0);
-    fill_viewport(t/100.0 * 0.25, 0, 0.8, "stage5/spell_clouds");
-    r_mat_mv_pop();
+	r_mat_mv_push();
+	r_mat_mv_translate(0, -100, 0);
+	fill_viewport(t/100.0, 0, 0.5, "stage5/spell_clouds");
+	r_mat_mv_translate(0, 100, 0);
+	fill_viewport(t/100.0 * 0.75, 0, 0.6, "stage5/spell_clouds");
+	r_mat_mv_translate(0, 100, 0);
+	fill_viewport(t/100.0 * 0.5, 0, 0.7, "stage5/spell_clouds");
+	r_mat_mv_translate(0, 100, 0);
+	fill_viewport(t/100.0 * 0.25, 0, 0.8, "stage5/spell_clouds");
+	r_mat_mv_pop();
 
-    float opacity = 0.05 * stage5_draw_data->stairs.light_strength;
-    r_color4(opacity, opacity, opacity, opacity);
-    fill_viewport(0, 300, 1, "stage5/spell_lightning");
+	float opacity = 0.05 * stage5_draw_data->stairs.light_strength;
+	r_color4(opacity, opacity, opacity, opacity);
+	fill_viewport(0, 300, 1, "stage5/spell_lightning");
 }
