+++ src/stages/stage5/meson.build
@@ -1,7 +1,16 @@
 stage5_src = files(
-    'background_anim.c',
-    'iku.c',
-    'draw.c',
-    'stage5.c',
-    'timeline.c',
+	'background_anim.c',
+	'iku.c',
+	'draw.c',
+	'stage5.c',
+	'timeline.c',
+    'nonspells/boss_nonspell_1.c',
+    'nonspells/boss_nonspell_2.c',
+    'nonspells/boss_nonspell_3.c',
+    'nonspells/midboss_nonspell_1.c',
+	'spells/atmospheric_discharge.c',
+	'spells/artificial_lightning.c',
+	'spells/induction.c',
+	'spells/natural_cathode.c',
+	'spells/overload.c'
 )
